% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_loader.R
\name{chunk_loader}
\alias{chunk_loader}
\title{Function to load big data to SQL (chunk.loader())}
\usage{
chunk_loader(...)
}
\arguments{
\item{...}{Not used.}
}
\description{
\strong{!!!STOP!!! This function has been deprecated.} Please use
\code{rads::tsql_chunk_loader()} instead.
}
\section{Deprecation}{

Please use \code{rads::tsql_chunk_loader()} instead.
}

