% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R, R/counts.R
\name{counts}
\alias{counts}
\alias{counts}
\title{Importing health conditions from the Medicaid claims data}
\usage{
counts(df, group_var = c("agency_new"), agency = "both", unit = NULL,
  filter = "", startdate = NULL, enddate = NULL, birth = NULL, ...)

counts(df, group_var = c("agency_new"), agency = "both", unit = NULL,
  filter = "", startdate = NULL, enddate = NULL, birth = NULL, ...)
}
\arguments{
\item{df}{A data frame}

\item{group_var}{A set of variables to group counts by. Default is PH agency.}

\item{agency}{A string that restricts analysis to KCHa, SHA or both. Default
is to count across both agencies (note that an individual or housheold will 
be counted only once if they appear in both agencies or >1 group for the
given time period). Not case sensitive.}

\item{unit}{A named variable that determines the unit to count over. 
Default unit of analysis is pid (individuals) but pid2 should be used with 
Medicaid-matched data. The other option is hhold_id_new for households.}

\item{filter}{A quosure that is a clunky way of adding the ability to add
extra filters. Default is missing. Addtional filters must be set up in a 
quo() syntax without quotation marks.}

\item{startdate}{A string that specifies the variable name for the time an
individual or household starts at that address/program. Default is set for
the linked PHA/Medicaid data (startdate_c) but if that is not present then
the next default is startdate.}

\item{enddate}{A string that specifies the variable name for the time an
individual or household ends participation at that address/program. 
Default is set for the linked PHA/Medicaid data (enddate_c) but if that is 
not present then the next default is enddate.}

\item{birth}{A named variable that determines which date of birth field to 
use when calculating age (normally dob_h or dob_m6 but can also use hh_dob_m6
if calculating age at the household level).}

\item{...}{Additional arguments for the \code{\link{time_range}} function
(yearmin, yearmax, period, and date).}

\item{df}{A data frame}

\item{group_var}{A set of variables to group counts by. Default is PH agency.}

\item{agency}{A string that restricts analysis to KCHA, SHA or both. Default
is to count across both agencies (note that an individual or housheold will 
be counted only once if they appear in both agencies or >1 group for the
given time period). Not case sensitive.}

\item{unit}{A named variable that determines the unit to count over. 
Default unit of analysis is pid (individuals) but pid2 should be used with 
Medicaid-matched data. The other option is hh_id_new for households.}

\item{filter}{A quosure that is a clunky way of adding the ability to add
extra filters. Default is missing. Addtional filters must be set up in a 
quo() syntax without quotation marks.}

\item{startdate}{A string that specifies the variable name for the time an
individual or household starts at that address/program. Default is set for
the linked PHA/Medicaid data (startdate_c) but if that is not present then
the next default is startdate.}

\item{enddate}{A string that specifies the variable name for the time an
individual or household ends participation at that address/program. 
Default is set for the linked PHA/Medicaid data (enddate_c) but if that is 
not present then the next default is enddate.}

\item{birth}{A named variable that determines which date of birth field to 
use when calculating age (normally dob_h or dob_m6 but can also use hh_dob_m6
if calculating age at the household level).}

\item{...}{Additional arguments for the \code{\link{time_range}} function
(yearmin, yearmax, period, and date).}
}
\description{
\code{conditions} brings in conditions from the Medicaid data.

\code{counts} summarizes the number of enrollees in the PHA data.
}
\details{
This function is a way to centralize defining and importing conditions from
the Medicaid claims data. This code will eventually be spun off into a 
Medicaid analysis package.

This function takes the date range supplied by \code{\link{time_range}} and
counts the number of individuals or households enrolled. Grouping options
are available and set using \code{\link{group_vars}}.
The function works for PHA data regardless of whether or not it has also
been matched to Medicaid data.
}
\examples{
\dontrun{
counts(pha_longitudinal)
counts(pha_longitudinal, group_var = c("agency_new", "major_prog"),
agency = "kcha", unit = hhold_id_new)
counts(pha_longitudinal, filter = quo(disability == 1))
}

\dontrun{
counts(pha_longitudinal)
counts(pha_longitudinal, group_var = c("agency_new", "major_prog"),
agency = "kcha", unit = hh_id_new)
counts(pha_longitudinal, filter = quo(disability == 1), yearmin = 2014,
yearmax = 2016, period = "month")
}

}
