% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumul_overlap.R
\name{cumul_overlap}
\alias{cumul_overlap}
\title{Counts the number of overlapping rows.}
\usage{
cumul_overlap(x)
}
\arguments{
\item{x}{The variable to count overlaps by (usually will be overlap)}
}
\description{
\code{cumul_overlap} counts the number of overlapping rows.
}
\details{
This function builds on \code{\link{overlap}} and counts the number of 
overlapping rows. Used in the \code{\link{los}} function.
Based on https://stackoverflow.com/questions/5012516/count-how-many-consecutive-values-are-true
}
\examples{
\dontrun{
cumul_overlap(overlap)
}
}
