% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatting.R
\name{format}
\alias{format}
\alias{date_ymd_f}
\alias{yesno_f}
\alias{char_f}
\alias{trim_f}
\title{General reformatting.}
\usage{
date_ymd_f(df, ...)

yesno_f(df, ...)

char_f(df, ...)

trim_f(df, ...)
}
\arguments{
\item{df}{A data frame}

\item{...}{Variables that need to be formatted}
}
\description{
Various functions that reformat variables as needed. These are
specific to the public housing authority datasets.
}
