% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_data.R
\name{get_raw_data}
\alias{get_raw_data}
\title{get_raw_data Gets raw nominal house price data from LU Open Data Portal}
\usage{
get_raw_data(
  url = "https://data.public.lu/fr/datasets/r/14b0156e-ff87-4a36-a867-933fc9a6f903"
)
}
\arguments{
\item{url}{Optional: Persistent url to the data}
}
\value{
A data frame
}
\description{
get_raw_data Gets raw nominal house price data from LU Open Data Portal
}
