% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housing-package.R
\docType{package}
\name{housing-package}
\alias{housing}
\alias{housing-package}
\title{housing: Housing Data For Luxembourg}
\description{
This package contains functions to get, clean and analyse housing price data for Luxembourg.
}
\author{
\strong{Maintainer}: Bruno Rodrigues \email{bruno@brodrigues.co} (\href{https://orcid.org/0000-0002-3211-3689}{ORCID})

}
\keyword{internal}
