% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{housing_default_options}
\alias{housing_default_options}
\title{Default options.}
\format{An object of class \code{list} of length 3.}
\usage{
housing_default_options
}
\description{
\code{apde_vars} sets default options for housing code specific to PHSKC use.
}
\details{
This function sets up the default options for working with the PHA data
while at PHSKC. Not needed for general use.
Also not working as desired. Setting global values in other R scripts for now
}
\keyword{datasets}
