% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/los.R
\name{los}
\alias{los}
\title{Sets up length of stay.}
\usage{
los(df)
}
\arguments{
\item{df}{A data frame}
}
\description{
\code{los} creates a move in date for calculating time in public housing.
}
\details{
This function takes the cleaned up/combined PHA data and calculates a set
of move-in dates. These can be used to calculate the length of time in 
housing overall (if the person moved between KCHA and SHA), at the
PHA, and in a given program.
Relies on \code{\link{overlap}} and \code{\link{cumul_overlap}}
For now, it is assumed that the df has variables called pid, period, 
agency_new, and major_program.
}
\examples{
#' \dontrun{
los(pha_longitudinal)
}
}
