% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{overlap}
\alias{overlap}
\title{Flags a row as overlapping with the previous row or not.}
\usage{
overlap(df, ...)
}
\arguments{
\item{df}{A data frame}

\item{...}{Arguments used to identify elements to check for matches}
}
\description{
\code{overlap} creates a flag indicating this row matches the previous one.
}
\details{
This function checks to see whether elements of this row match the previous
row's elements. Its output is a new variable in the data frame. The main 
purpose is for use in the \code{\link{los}} function.
}
\examples{
\dontrun{
overlap(pha_longitudinal, pid, agency_new)
}
}
