% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popcount.R
\name{popcount}
\alias{popcount}
\title{Counting population and person-time in the joint PHA/Medicaid data}
\usage{
popcount(df, group_var = quos(agency_new), agency = NULL, enroll = NULL,
  unit = NULL, startdate = NULL, enddate = NULL, birth = NULL,
  length = NULL, numeric = TRUE, wc = FALSE, ...)
}
\arguments{
\item{df}{A data frame}

\item{group_var}{A set of variables to group counts by. Default is PH agency.
Must be set as a quosure (use quos(<group1>, <group2>))}

\item{agency}{A named variable that specifies the agency a person is in for 
that period of time (usually KCHA, SHA, or NA/Medicaid only). Used to 
allocate individuals who moved between multiple agencies/enrollment types 
in the period. Default is agency_new.}

\item{enroll}{A named variable that specifies the type of enrollment a person
is in for that period of time (field should contain the following codes: 
"m" = Medicaid only, "b" = housing and Medicaid, or "h" = housing only).
Used to allocate individuals who moved between 
multiple agencies/enrollment types in the period. Default is enroll_type.}

\item{unit}{A named variable that determines the unit to count over. 
Default unit of analysis is pid2 (individuals) but pid should be used with 
data not matched to Medicaid. The other option is hhold_id_new for households.}

\item{startdate}{A string that specifies the variable name for the time an
individual or household starts at that address/program. Default is set for
the linked PHA/Medicaid data (startdate_c) but if that is not present then
the next default is startdate.}

\item{enddate}{A string that specifies the variable name for the time an
individual or household ends participation at that address/program. 
Default is set for the linked PHA/Medicaid data (enddate_c) but if that is 
not present then the next default is enddate.}

\item{birth}{A named variable that determines which date of birth field to 
use when calculating age (normally dob_c or dob_h but can also use 
hh_dob_m6 if calculating age at the household level). Only needed if using an
age grouping or the well-child flag.}

\item{length}{A named variable that determines which start date to 
use when calculating length of time in housing (normally start_housing).}

\item{numeric}{Determine whether to recode age and length of time groups to
be numeric or remain as character.}

\item{wc}{Choose whether to just run pop counts for the well-child indicator
age group (ages 3-6 years)}

\item{...}{Additional arguments for the \code{\link{time_range}} function
(yearmin, yearmax, period, and date).}
}
\description{
\code{popcount} summarizes population data in the joint PHA/Medicaid data.
}
\details{
This function takes the date range supplied by \code{\link{time_range}} and
counts individuals enrolled and person time by group.
The function works for PHA data regardless of whether or not it has also
been matched to Medicaid data.
}
\examples{
\dontrun{
popcount(pha_longitudinal)
popcount(pha_longitudinal, group_var = quos(agency_new, major_prog),
agency = "kcha", unit = hhold_id_new)
popcount(pha_longitudinal, yearmin = 2014, yearmax = 2016, period = "month")
}

}
