% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streets.R
\name{streets}
\alias{streets}
\title{Standardizes street names.}
\usage{
streets(df)
}
\arguments{
\item{df}{A data frame}

\item{...}{A set of variables names that need cleaning}
}
\description{
\code{streets} standardizes street names to have USPS abbreviations.
}
\details{
This function applies standard USPS abbreviations to variables containing
street names
}
\examples{
\dontrun{
los(pha_longitudinal, unit_add_new)
}

}
