% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_range.R
\name{time_range}
\alias{time_range}
\title{Set up time periods.}
\usage{
time_range(yearmin = 2012, yearmax = 2016, period = "year",
  date = "-12-31")
}
\arguments{
\item{yearmin}{An integer that sets the starting point for the time range.}

\item{yearmax}{An integer that sets the end point for the time range.}

\item{period}{A string of either date, month, quarter, or year the determines
the length of time to count over. Default is year.}

\item{date}{A string that can be used to set the specific date to look at.
Must be used with period set to date. Takes the format "-MM-DD". Non-valid
dates will produce a charToDate(x) error.}
}
\description{
\code{time_range} creates vectors of start and end dates to count over
}
\details{
This function sets up the vectors for which we want to count PHA enrollment.
Its output is a list of start and end vectors, plus a string showing the 
period of time being counted in each start/end combo.
}
\examples{
\dontrun{
time_range()
time_range(yearmin = 2014, yearmax = 2015, period = "quarter")
time_range(period = "date", date = "-06-30")

time_range(yearmin = 2017, yearmax = 2016)
}

}
