% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yesler.R
\name{yesler}
\alias{yesler}
\title{Flags people who live at Yesler Terrace.}
\usage{
yesler(df, unit = NULL)
}
\arguments{
\item{df}{A data frame}

\item{unit}{A named variable that determines the unit to group by. Default is 
pid2 (individuals) but pid should be used with housing-only data that has not
been matched to Medicaid data.}
}
\description{
\code{yesler} creates a set of flags specific to SHA's Yesler Terrace.
}
\details{
This function creates a set of flags that are specific to Seattle Housing
Authority's Yesler Terrace site. It is a convenient way to quickly add the 
flags to a given data set. Also adds a flag for scattered sites.
Once each person's addresses have been flagged as Yesler Terrace/scattered 
sites or not, a new set of variables is made to indicate whether or not the 
person has ever lived at Yesler Terrace/scattered sites.
For now, it is assumed that the df has variables called property_id, unit_add_new, 
and variable to group people on (pid or pid2).
}
\examples{
#' \dontrun{
yesler(pha_elig_demogs, unit = pid)
}

}
