% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yt_flag.R
\name{yt_flag}
\alias{yt_flag}
\title{Flags people who live at Yesler Terrace.}
\usage{
yt_flag(df, unit = NULL, prop_id = NULL, pop_name = NULL,
  address = NULL)
}
\arguments{
\item{df}{A data frame}

\item{unit}{A named variable that determines the unit to group by. Default is 
pid2 (individuals) but pid should be used with housing-only data that has not
been matched to Medicaid data.}

\item{prop_id}{A named variable that specifies the property IDs used in 
housing data. Default is property_id.}

\item{address}{A named varaible that specifies the unit address. 
Default is unit_add_h.}

\item{prop_id}{A named variable that specifies the property name used in 
housing data. Default is property_name.}
}
\description{
\code{yt_flag} creates a set of flags specific to SHA's Yesler Terrace.
}
\details{
This function creates a set of flags that are specific to Seattle Housing
Authority's Yesler Terrace site. It is a convenient way to quickly add the 
flags to a given data set. Also adds a flag for scattered sites.
Once each person's addresses have been flagged as Yesler Terrace/scattered 
sites or not, a new set of variables is made to indicate whether or not the 
person has ever lived at Yesler Terrace/scattered sites.
It is assumed that the df has variables for the property ID, property name, 
address, and individual identifier (pid or pid2).
}
\examples{
\dontrun{
yt_flag(pha_elig_demogs, unit = pid)
}

}
