% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yt_popcode.R
\name{yt_popcode}
\alias{yt_popcode}
\title{Assign people to a demographic group for a given year, prioritize YT.}
\usage{
yt_popcode(df, year = 12, year_pre = "pt", year_suf = NULL,
  agency = NULL, enroll_type = NULL, dual = NULL, yt = NULL,
  ss = NULL, pt_cut = 30, min = T)
}
\arguments{
\item{df}{A data frame}

\item{year}{An integer that describes the calendar year being examined. 
Used to identify the person-time variable.}

\item{year_pre}{A string that prefixes the year in the person-time variable.}

\item{year_suf}{A string that suffixes the year in the person-time variable.}

\item{agency}{A named variable that specifies the agency.}

\item{enroll_type}{A named variable that specifies the enrollment type.}

\item{dual}{A named variable that specifies a person's dual eligibility.}

\item{yt}{A named variable that specifies the Yesler Terrace flag. Must be 
binary (0/1).}

\item{ss}{A named variable that specifies the scattered sites flag. Must be
binary (0/1).}

\item{pt_cut}{The threshold in person-days a person should meet in order to 
be considered at YT/SS in that year}

\item{min}{Flag to keep the minimum pop_code.}
}
\description{
\code{yt_popcode} assigns people in the SHA data to a demographic group.
}
\details{
This function is specific to Seattle Housing Authority's data. It assigns a 
numeric code to each row in the data where person-time has accumulated for 
that calendar year. Codes are assigned as follows:
(#1-5 assume simultaneous Medicaid enrollment and not dual eligible):
1) At least 30 days at YT that year = YT, 
2) No YT but 30+ days at SS that year = SS, 
3) No YT or SS but in SHA for 30+ days = SHA, 
4) Not YT/SS/SHA but in KCHA for 30+ days = KCHA, 
5) No 30+ days in any PHA = non-PHA Medicaid, 
6-10) Same as #1-5 but only dual eligible, 
11-15) Same as #1-5 but dual eligible is NA, 
16-18) Only time in SHA (+/- YT and SS) and no Medicaid, 
19) Only time in KCHA and no Medicaid

If the min flag is true, only takes the smallest numeric code produced is 
kept for that year and new demographic variables are calculated. 
Taking the smallest code prioritizes a person's time at Yesler Terrace when 
they were also enrolled in Medicaid.

It is assumed that the df has variables for housing authority, enrollment 
type, dual eligibility, and person-time for that calendar year. Also assumes
that \code{\link{yt_flag}} has been run to identify people living at YT
or scattered sites.

Eventually could expand to accommodate non-calendar years using the 
\code{\link{time_range}} function.
}
\examples{
\dontrun{
yt_popcode(yt_mcaid_final, year_pre = "pt", year = 12, min = T)
}

}
