% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junk_ssn.R
\name{junk_ssn}
\alias{junk_ssn}
\alias{junk_ssn_num}
\alias{junk_ssn_char}
\alias{junk_ssn_all}
\title{Makes a flag to indicate whether or not a SSN is junk}
\usage{
junk_ssn_num(df, id = NULL)

junk_ssn_char(df, id = NULL)

junk_ssn_all(df, id = NULL)
}
\arguments{
\item{df}{A data frame}

\item{id}{Name of variable to check for accuracy of a SSN or PHA ID}
}
\description{
\code{junk_ssn} makes a flag to indicate whether or not a SSN is junk.
}
\details{
This function checks to see whether the Social Security Number provided could
be legitimate or not. See here for guidance: 
https://www.ssa.gov/employer/verifySSN.htm
The function also checks whether or not a housing ID is likely to be a 
temporary designation that is shared by many others.
}
